+++ src/components/reducer.ts
@@ -7,7 +7,7 @@ import { clearProperties, getMethods, isPrimitive, log, simpleCombineReducers } 
 import { ComponentActions, ReduxAppAction } from './actions';
 import { Component } from './component';
 
-// tslint:disable:member-ordering
+// tslint:disable:member-ordering ban-types
 
 export type ReducerCreator = (changeListener: Listener<Component>) => Reducer<object>;
 
@@ -39,33 +39,17 @@ export class ComponentReducer {
 
     public static createReducer(component: Component, componentCreator: object): ReducerCreator {
 
-        // method names lookup
-        const methods = getCreatorMethods(componentCreator);
         const creatorInfo = CreatorInfo.getInfo(componentCreator);
         if (!creatorInfo)
             throw new Error(`Inconsistent component '${componentCreator.constructor.name}'. The 'component' class decorator is missing.`);
 
-        const options = creatorInfo.options;
-        const methodNames: any = {};
-        Object.keys(methods).forEach(methName => {
-
-            // reducers does not handle 'noDispatch' and 'sequence' methods
-            if (creatorInfo.noDispatch[methName] || creatorInfo.sequence[methName])
-                return;
-                
-            var actionName = ComponentActions.getActionName(componentCreator, methName, options);
-            methodNames[actionName] = methName;
-        });
-
-        // state object prototype
+        const methods = ComponentReducer.createMethodsLookup(componentCreator, creatorInfo);
         const stateProto = ComponentReducer.createStateObjectPrototype(component, creatorInfo);
-
-        // component id
         const componentId = ComponentInfo.getInfo(component).id;
-
-        // the reducer
+        
         return (changeListener: Listener<Component>) => {
 
+            // the reducer
             return (state: object, action: ReduxAppAction) => {
 
                 log.verbose(`[reducer] Reducer of: ${componentCreator.constructor.name}, action: ${action.type}`);
@@ -83,8 +67,7 @@ export class ComponentReducer {
                 }
 
                 // check if should use this reducer
-                const methodName = methodNames[action.type];
-                const actionReducer = methods[methodName];
+                const actionReducer = methods[action.type];
                 if (!actionReducer) {
                     log.verbose('[reducer] No matching action in this reducer, returning previous state');
                     return state;
@@ -112,8 +95,9 @@ export class ComponentReducer {
             const start = Date.now();
 
             context.invoked = true;
+            log.debug(`[rootReducer] Reducing action: ${action.type}.`);
 
-            var newState = reducer(state, action);
+            let newState = reducer(state, action);
             newState = ComponentReducer.finalizeState(newState, root);
 
             const end = Date.now();
@@ -126,17 +110,24 @@ export class ComponentReducer {
     //
     // private methods - state object
     //
+    
+    private static createMethodsLookup(componentCreator: object, creatorInfo: CreatorInfo): IMap<Function> {
+        
+        const allMethods = getCreatorMethods(componentCreator);        
 
-    /**
-     * Create a "state object". The state object receives it's properties from
-     * the current state and it's methods from the owning component. Methods
-     * that represent actions are replace with a throw call, while noDispatch
-     * methods are kept in place.
-     */
-    private static createStateObject(state: object, stateProto: object): object {
-        const stateObj = Object.create(stateProto);
-        Object.assign(stateObj, state);
-        return stateObj;
+        const options = creatorInfo.options;
+        const actionMethods: IMap<Function> = {};
+        Object.keys(allMethods).forEach(methName => {
+
+            // reducers does not handle 'noDispatch' and 'sequence' methods
+            if (creatorInfo.noDispatch[methName] || creatorInfo.sequence[methName])
+                return;
+
+            var actionName = ComponentActions.getActionName(componentCreator, methName, options);
+            actionMethods[actionName] = allMethods[methName];
+        });
+
+        return actionMethods;
     }
 
     /**
@@ -159,6 +150,18 @@ export class ComponentReducer {
         throw new Error("Only 'noDispatch' methods can be invoked inside actions.");
     }
 
+    /**
+     * Create a "state object". The state object receives it's properties from
+     * the current state and it's methods from the owning component. Methods
+     * that represent actions are replace with a throw call, while noDispatch
+     * methods are kept in place.
+     */
+    private static createStateObject(state: object, stateProto: object): object {
+        const stateObj = Object.create(stateProto);
+        Object.assign(stateObj, state);
+        return stateObj;
+    }
+
     //
     // private methods - combine reducers
     //
@@ -179,7 +182,7 @@ export class ComponentReducer {
             return ComponentReducer.identityReducer;
 
         // get the root reducer
-        var rootReducer: Reducer<any>;
+        let rootReducer: Reducer<any>;
         const info = ComponentInfo.getInfo(obj as any);
         if (info) {
             rootReducer = info.reducerCreator(comp => {
@@ -198,7 +201,7 @@ export class ComponentReducer {
                 continue;
 
             // other objects
-            var newSubReducer = ComponentReducer.combineReducersRecursion((obj as any)[key], new CombineReducersContext({
+            const newSubReducer = ComponentReducer.combineReducersRecursion((obj as any)[key], new CombineReducersContext({
                 ...context,
                 path: (context.path === '' ? key : context.path + '.' + key)
             }));
@@ -206,11 +209,11 @@ export class ComponentReducer {
                 subReducers[key] = newSubReducer;
         }
 
-        var resultReducer = rootReducer;
+        let resultReducer = rootReducer;
 
         // combine reducers
         if (Object.keys(subReducers).length) {
-            var combinedSubReducer = simpleCombineReducers(subReducers);
+            const combinedSubReducer = simpleCombineReducers(subReducers);
 
             resultReducer = (state: object, action: ReduxAppAction) => {