+++ src/index.ts
@@ -2,3 +2,4 @@ export { isInstanceOf } from './components';
 export { component, connect, computed, ignoreState, noDispatch, sequence, withId } from './decorators';
 export { SchemaOptions, ComputedOptions, AppOptions, GlobalOptions, LogLevel } from './options';
 export { ReduxApp } from './reduxApp';
+export { getMethods } from './utils';