+++ src/utils/utils.ts
@@ -15,7 +15,7 @@ export function isPrimitive(val: any): boolean {
     return type !== 'object' && type !== 'function';
 }
 
-export function getMethods(obj: object | Function): IMap<Method> {
+export function getMethods(obj: object | Function, bind = false): IMap<Method> {
     if (!obj)
         return undefined;
 
@@ -38,8 +38,12 @@ export function getMethods(obj: object | Function): IMap<Method> {
         var desc = Object.getOwnPropertyDescriptor(proto, key);
         var hasGetter = desc && typeof desc.get === 'function';
 
-        if (!hasGetter && typeof proto[key] === 'function')
+        if (!hasGetter && typeof proto[key] === 'function') {
             methods[key] = proto[key];
+            if (bind) {
+                methods[key] = methods[key].bind(obj);
+            }
+        }
     }
 
     return methods;
@@ -86,6 +90,33 @@ export function getParentType(obj: object | Function) {
 
     // get parent type
     return Object.getPrototypeOf(type.prototype).constructor;
+}
+
+/**
+ * Test if an object is a plain object, i.e. is constructed by the built-in
+ * Object constructor and inherits directly from Object.prototype or null. 
+ *
+ * Some built-in objects pass the test, e.g. Math which is a plain object and
+ * some host or exotic objects may pass also.
+ * 
+ * https://stackoverflow.com/questions/5876332/how-can-i-differentiate-between-an-object-literal-other-javascript-objects
+ */
+export function isPlainObject(obj: any) {
+    if (!obj)
+        return false;
+
+    if (typeof obj !== 'object')
+        return false;
+
+    // if Object.getPrototypeOf supported, use it
+    if (typeof Object.getPrototypeOf === 'function') {
+        var proto = Object.getPrototypeOf(obj);
+        return proto === Object.prototype || proto === null;
+    }
+
+    // otherwise, use internal class
+    // this should be reliable as if getPrototypeOf not supported, is pre-ES5
+    return Object.prototype.toString.call(obj) === '[object Object]';
 }
 
 export function toPlainObject(obj: any): any {